+++ CHANGELOG.md
@@ -11,9 +11,21 @@ We do not give estimated times for completion on `Accepted` Proposals.
 
 ---
 
+### v6.0.1 - 2017-02-25
+
+`FIXED`
+
+- A few config references.
+
 ### v6.0.0 - 2017-02-24
 
 - Updated for Platform 7.
+
+### v5.0.1 - 2017-02-25
+
+`FIXED`
+
+- A few config references.
 
 ### v5.0.0 - 2017-02-24